Don't reuse flags from the crate when generating its bindings.
The bindings are independent -- for example, the crate might forbid unsafe code, but the bindings, being something with FFI, must allow it.

---

Consider the following Rust library:

```rust
#[cfg(define_foo)]
pub fn foo() -> X {...}
```

With the following BUILD configuration:

```python
    rustc_flags=["--cfg=define_foo", "-Dunsafe-code"],
```

When we build the bindings for this, we would generate something like the following thunk:

```rust
pub extern "C" fn foo(output: *mut X) {
  unsafe {*output = real_library::foo();}
}
```

It is _not_ correct to apply the rustc flags from the library onto the thunk. It is at best not helpful: passing in `--cfg=define_foo` doesn't change the correctness of the thunk. It is at worst actively harmful: passing in `-Dunsafe-code` causes compilation of the thunk to fail entirely.

Like any other user of a library, the flags passed to the thunk library ought to be independent of the flags passed to the library it depends on.

PiperOrigin-RevId: 627499845
Change-Id: I206a5e4ebda0e529580182c101d2d0885d9bd62f