Move crubit underneath the rules_rust process_wrapper.
(This fixes the last remaining known failures for Crubit on third-party crates.)

This means it effectively "just" pretends to be rustc, with some extra flags at the start. Any preprocessing the process wrapper does, like unpacking arguments from `build.rs` files, is done for us with code shared with normal `rustc` invocations.

This solves a number of correctness issues all at once, and in a generalized way that means we don't have to keep spending our time patching these problems. It also lets us remove code where we duplicate process wrapper logic, and means we encode fewer bazel-specific assumptions in the tool binary itself.

It's still possible that args-forwarding could break in the future, like if there were another field added to `args`, but this seems unlikely. (If it does happen, we can take the pains to switch to `.all`.)

A note: this CL works even without any changes to cc_bindings_from_rs. It's just that the code in cc_bindings_from_rs.rs, after this CL, is dead code, so it can be cleaned up all in one fell swoop.

I am very happy with this CL! Huge amount of bang-for-buck here.

PiperOrigin-RevId: 627893862
Change-Id: I0d3dc4a1e7e5d7ccc3b68cf0a5bf0fde6b054fa3