Collect evidence from global variable initializers.
Replaces use of `Implementation` with `Definition` to more closely describe the expanded set of Decls processed by this collection function. While updating comments in collect_evidence, change decl-level comments to use Doxygen-style `///`.

We may be able to use a similar CFG-from-Expr pattern to collect fully-analyzed evidence for default arguments and fields for types that don't have a defined constructor.

PiperOrigin-RevId: 629420396
Change-Id: Ib03783ba7b8c407747bd7fff62a30d5aeacc913c