Add support for #[repr(C)] union fields
Added basic support for unions that follow the C memory layout. We basically do the same thing as generating bindings for structs except we don't wrap the fields, we provide the correct size and alignment, and we don't allow the packing attribute to be used.

Because unions are technically a special type of class (with constructors and destructors), the logic for generating struct constructors/destructors is used here.

Information on unions - https://en.cppreference.com/w/cpp/language/union

PiperOrigin-RevId: 631055981
Change-Id: I02d0b3adad44d42897dc82fd51764c9e030e9221