Enable smart pointers in inference tests that will be affected by smart pointer support.
This enables analysis of smart pointers, but not evidence collection for them. We can make use of the analysis to collect additional evidence for raw pointers, demonstrated by the added test.

I didn't expect smart pointer analysis to break any existing tests for raw pointers or for changes to smart pointer analysis to break them, but now we'll know if they do. This also sets a baseline as I start to add smart pointer tests and functionality for inference.

PiperOrigin-RevId: 631427810
Change-Id: I077477946d52f5eed725368234d5f4622aba563d