Use a slightly different Environment API.
This maintains the assertion that Env's current analysis target is a FunctionDecl and is not null, but is also compatible with https://github.com/llvm/llvm-project/pull/91616. It does not require that PR to work though, which means we don't need this to be managed as part of the integrate.

PiperOrigin-RevId: 632566151
Change-Id: I8930da018d860a4902a984822c6cac888a065d17