Add support for rust union fields
Added basic support for unions that follow the Rust memory layout. Basically, we wrap the union fields in a struct that contains the acceptable padding so it can be accessed. Note this is not done for unknown fields (which are marked private).

Also added tests.

Note Rust doesn't always guarantee that union fields will have a 0 offset from the beginning of the union: https://rust-lang.github.io/unsafe-code-guidelines/layout/unions.html

See https://doc.rust-lang.org/beta/nightly-rustc/src/rustc_abi/lib.rs.html#1186-1203. We depend on this function to determine the offset (which right now will be zero, but may change).

PiperOrigin-RevId: 633587953
Change-Id: Icc2e9466fdd8b9843556894e83c0ef283623008b