Collect inference evidence for Nonnull-ness from nonnull attribute.
Using a soft rule only because if e.g. someone were somehow getting past the compiler check and assigning a nullable, safety would be violated by applying a Nonnull inference instead of a Nullable one. Not treating it the same as one of our annotations because verification won't be reading the attribute as an annotation or checking for conflicts between ours and these attributes.

PiperOrigin-RevId: 641892104
Change-Id: Ie51a6f624b6eef643d104e24d22828243fec539c