Roll forward smart pointer evidence from constructor exit blocks with fix.
Fix is to return early in this case where the null state is not present. This querying of null state is special in that it is not for a visited expression, which means the null state for the relevant smart pointer field may not have been attached because it was copied or moved from a field that was not modeled.

Also add a more direct CHECK for null state presence in other cases.

PiperOrigin-RevId: 641941553
Change-Id: I407cafc4717767214d2142ffe2b357508623671b