Store and retrieve canonical decls in nullability overrides.
Affects symbolic overrides for SAT queries and also concrete overrides for previous inferences. In both cases, this shouldn't have been incorrect before because of how the Decls we look up are related to the Decls we override and because the fingerprinting for previous inferences should produce the same fingerprints for canonical and non-canonical Decls. But better resilient than sorry.

PiperOrigin-RevId: 642691739
Change-Id: Ib03fa204b87a47d3b665b22f5a23c7a5efa051ab