Expose API for getting edit ranges that are not inference targets.
This can be used for adding/removing explicit annotations when adding a default nullability setting, for ranges that are not (yet) targeted by inference tooling.

Renames existing API to clarify the distinction.

No test included, because it would be sensitive only to action at a distance when the definition of inference targets change.

PiperOrigin-RevId: 643984048
Change-Id: I8f78241b239a4dc0e216ebac5a573a7df4bdd356