Teach pragma about transparent aliases and reduce scope and effect of Unknown annotation.
---

In general, the governing pragma is determined by the least-sugared TypeLoc:
- for raw pointers, the PointerType (where the star is)
- for smart pointers, the TemplateSpecializationType for the class template

This patch introduces an exception: in the case of "transparent alias" sugar,
we prefer the sugared form.
The motivating case is:

    template <class T> class MyLongPointer {};
    // Give MyLongPointer a shorter name.
    template <class T> using MyPtr = MyLongPointer<T>;

When code refers to MyPtr<T>, we want to consider the pragma at the use site,
not in the file where the MyPtr alias is defined.
For now, a "transparent alias" is an alias template which forwards its template
parameters exactly into a target template.

---

Until now we have treated "Unknown" as a regular nullability annotation, i.e.
a strong assertion about nullability that should override others.

This patch changes its meaning to "disregard this file's pragma for this type".
This is more useful for incremental migration.

Unknown can now only be applied directly to a supported pointer type: if it
is applied to a type alias, it will be ignored.

PiperOrigin-RevId: 643997928
Change-Id: Icf0174715f57eaea743718a14e92a74000d82e3b