Enable removal of Unknown annotations when Nonnull/Nullable inferences are produced.
Only looks for absl::NullabilityUnknown<...> and _Null_unspecified as an Unknown annotation to be removed.

If we have determined that there is an Unknown annotation and a Nonnull/Nullable inference, but we don't have lengths, we leave the Unknown annotation in place to avoid producing useless edits that may not build or be readable by verification tooling.

PiperOrigin-RevId: 645390584
Change-Id: Idae050a168f0902df3871e612df561fbe6e84281