[nullability] Split diagnosis into two callbacks.
The framework recently added the ability to run post-analysis callbacks not just
on the state after the transfer function for a `CFGElement` has been run but
also before it.

Conceptually, diagnostic callbacks usually make more sense to run on the
"before" state because we usually want to check preconditions for the operation,
so almost all diagnostics are now run in this way.

This makes the diagnostic callback for smart pointers more intuitive: When we
were running this on the "after" state, to check whether we were assigning a
nullable pointer to a nonnull variable, we somewhat unintuitively had to check
the nullability of the _left-hand_ side (because the right-hand side of a smart
pointer assignment is always null after the assignment has been carried out).
Now that the callback is running on the "before" state, we can more intuitively
check the nullability of the right-hand side.

More importantly, this will enable us to perform diagnosis correctly on the `++`
and `--` operators. For these, we need to access the "before" state to check
whether the pointer being incremented is nullable. (The "after" state of the
pointer is always non-null.)

There is one check that we are still running on the "after" state, namely
`diagnoseMovedFromNonnullSmartPointer`. This needs to be run on the "after"
state because in the "before" state, the pointer and its nullability properties
may not yet be initialized.

PiperOrigin-RevId: 646351557
Change-Id: Ib51e4d236bfc8662f55ef3e6537d65bb1a9d0c57