[nullability] Rework the way we ensure raw pointers are initialized.
We’re making this more consistent with smart pointers in that we give our
transfer functions a chance to initialize the pointer first and only create a
“backup” value if the transfer functions didn’t initialize the pointer.

I was prompted to make this change when I observed that, for raw pointer
glvalues, `ensurePointerHasValue()` would create a storage location if necessary
but not associate it with the expression. This seemed like a simple fix, but
merely adding an `Env.setStorageLocation()` caused failures because
`modelGetReferenceableValue()` was now trying to set a storage location for an
expression that already had one.

This can be fixed by moving `ensurePointerHasValue()` until after the transfer
function has run, which makes things more consistent with the treatment of smart
pointers too. To emphasize the distinction between raw and smart pointers, I am
renaming `ensurePointerHasValue()` to `ensureRawPointerHasValue()`.

Some transfer functions previously assumed that the expression they operate on
was already guaranteed to be associated with a pointer. Now, transfer functions
need to take care of this themselves if necessary. To do this, I am providing an
overload of `ensureRawPointerHasValue()` that takes an expression and returns
the pointer associated with that expression (either an already existing pointer
or a newly created one).

PiperOrigin-RevId: 646927739
Change-Id: Id05e9fff268127814606b97dadd7cd9c995eed97