Use standard variadic macros instead of the GNU extension in `CRUBIT_OFFSETOF`.
Right now, we seem to use `CRUBIT_OFFSETOF` from generated files and then build them with the flags from the target the bindings are for, which is probably wrong. This surfaces as parsing failures, when they enable additional warnings. In this case, the warning is correct, I think: we should use the standard va_args mechanism, not a GNU preprocessor extension.

This is one of multiple errors that surface in b/348557947, but doesn't really fix the root cause of that bug, or even of this compilation error. Still, it's an improvement anyway.

PiperOrigin-RevId: 647809127
Change-Id: I31673d34b5b262ab5d65f0202b9a453be0a98f08