Disable Crubit on grpc and targets that disable header parsing.
This allows Crubit to work on targets that depend on grpc, as long as they do not transitively include the headers. So, that's a workaround, at least.

This is not "strictly correct": a target might disable header parsing because it has a bad header in `srcs`, which isn't relevant to Crubit which only reads `hdrs`. However, it's almost certainly (also) broken in `srcs`. If ever we need to enable Crubit on a target that has a broken srcs header, but not a broken.

As far as grpc itself goes, we just don't correctly handle those headers right now. If grpc shows up anywhere in the transitive dependency graph, it breaks Crubit. That's a problem, and this works around that by disabling Crubit.

(Similarly for openssl.)

In the longer term, we should not have the grpc denylist, and I've filed b/349750606 and b/349750659 to track some of this work.

PiperOrigin-RevId: 647810496
Change-Id: Ic75dd706304d4373afa6c565633dd9c397b042cc