Automated rollback of commit 240782f5d5cae15c48fb55d9ca68e02ea68cf3ee.
*** Reason for rollback ***

Break cafe.bow.asan.

*** Original change description ***

Disable Crubit on grpc and targets that disable header parsing.

This allows Crubit to work on targets that depend on grpc, as long as they do not transitively include the headers. So, that's a workaround, at least.

This is not "strictly correct": a target might disable header parsing because it has a bad header in `srcs`, which isn't relevant to Crubit which only reads `hdrs`. However, it's almost certainly (also) broken in `srcs`. If ever we need to enable Crubit on a target that...

***

PiperOrigin-RevId: 647840087
Change-Id: I847b2ea3444b8e3932c41fe3bd0f539fcc69f107