Rollback of:
Disable Crubit on grpc and targets that disable header parsing.

This allows Crubit to work on targets that depend on grpc, as long as they do not transitively include the headers. So, that's a workaround, at least.

This is not "strictly correct": a target might disable header parsing because it has a bad header in `srcs`, which isn't relevant to Crubit which only reads `hdrs`. However, it's almost certainly (also) broken in `srcs`. If ever we need to enable Crubit on a target that...

PiperOrigin-RevId: 648267674
Change-Id: I35cbdeed7928cfe93b133221b7201ddacb2bf813