Make Bazel tests for Crubit feature infra independent of production features
We update the definitions of feature sets like "supported" and "experimental"
for release engineering reasons. The expansion of these feature sets is
hardcoded in a Bazel test. As a result, when we update the feature sets to
release a feature or clean up a feature flag, we have to update a Bazel test.

This change makes Bazel tests use special test-only feature set names that are
stable, thus eliminating the churn in the Bazel test.

PiperOrigin-RevId: 650151945
Change-Id: I2620a2dde83fd341f06f03f20f2ea3818fe0852b