Move support for non-"extern C" functions to a new feature, `non_extern_c_functions`
I changed the feature check call `require_features` into `require_any_feature` to make `non_extern_c_function` be a member of the `experimental` feature set *internally in Crubit* and not just in Bazel. This way we don't have to plumb `non_extern_c_function` through a variety of places, like `extern_c` was. When we decide to stabilize `non_extern_c_function` we can make it a part of the `supported` by changing the definition of the aspect hint.

PiperOrigin-RevId: 650211217
Change-Id: Ib56aed008262ecb71a4eb9f00357ba27095adbb8