Produce ranges for nested array/pointer/function pointer types with complex declarators.
These types need special handling due to the placement of names and array brackets that is different when the pointer types are inside an alias template (such as a nullability annotation).

This won't cover all cases, but those not covered are even more rare than these complex declarators already are in general.

PiperOrigin-RevId: 651443512
Change-Id: I265849ba30466e8d90e1f8b7ecc274bb1a4d95cd