[nullability] Collect evidence from the state before transfer function ran.
As for diagnosis (see a [similar
change](https://github.com/google/crubit/commit/2495238521651a34b79f08c8e13ed457dbfb5474)
made there recently), it makes more sense to perform evidence collection before
the transfer function is run (instead of after), as we know what the
preconditions for the operation are and can therefore, from these preconditions,
collect evidence for what the input values of the operation should be.

This patch not only fixes a regression in
CollectEvidenceFromDefinitionTest.Arithmetic caused by a recent upstream change,
but it also fixes existing wrong behavior in
CollectEvidenceFromDefinitionTest.ReferenceArgsPassed.

It also makes evidence collection for assignment more natural (see the lengthy
comment that previously existed and that this patch now eliminates).

PiperOrigin-RevId: 652814773
Change-Id: Iedf97419beb60a5d628a381e0072370fd755df46