Fix b/333737712: incorrect namespace assignment for items in reopened namespaces.
This ended up being a pretty simple problem after all. We do have independent namespace decl IDs for the original vs reopened namespace, so there _is_ an item ID that the nested declaration could be assigned as a parent -- it just was assigned the wrong one. Why? Well, the previous version of the code used `getEnclosingNamespaceContext`, which loops over the parents until it finds the namespace, and then returns the _canonicalized_ namespace decl: the _first_ namespace decl with that name, not the actual direct parent, which might be a redeclaration. The fix, then, is to not use `getEnclosingNamespaceContext`, but to reproduce the same loop in our code.

PiperOrigin-RevId: 655052095
Change-Id: Ic2b8c85f65a8bdb6f4de55f61231bc5f3ef49e47