Status page: "are we Crubit yet?"
(The title is in the vein of https://wiki.mozilla.org/Areweyet. I didn't even know this had a wiki page tracking it...)

This might not be super useful externally, right now, because it's primarily a bag of links to bugs, but there's also nothing particularly secret here. I opted not to curate the github markdown or anything, but also not redact it from the Copybara process.

PiperOrigin-RevId: 657358728
Change-Id: I6a31675a3c8cfd4b69b5a159a9a0c852609e7982