Support `char` in public interfaces in `:supported`.
This was, as far as I know, an unintentional omission of https://github.com/google/crubit/commit/1dd44144c97223ae7602e0dcfc4e44eca43c2f8e. If it was intentional, well, I didn't document it well enough, so whatever.

Certainly there's not much harm in it -- yes, I _do_ want to replace `ffi::c_char` with e.g. an ffi newtype, but that is a separate effort. People need to be able to use `const char*`, and we settled on doing the "obvious" thing, the same as with ordinary `extern "C"` FFI, in the MVP.

The error was hidden somewhat by the other issues noted in b/356481403, which caused structs that use `char` to still work as expected.

PiperOrigin-RevId: 658137222
Change-Id: Ifd810c693d715ef36622588621b716936a854505