Support `-Cpanic=unwind` (on nightly), by just changing documentation.
Crubit is made sound for free by the changes in nightly to abort when panicking through `extern "C"`. These are documented here: https://doc.rust-lang.org/nomicon/ffi.html#ffi-and-unwinding -- however, note that this functionality has not yet hit Rust stable. The documentation change was made too soon, as far as I can tell. Still, this _will_ be the behavior, and when it is, we will be sound.

For `extern "C-unwind"`, I think our docs here were inaccurate, and it's not sound unless the caller enables exceptions. I suspect we should not release this functionality to end-users yet, as it's dangerous. Maybe we should instead default to the sound option of aborting.

We probably shouldn't say the issue of `panic=unwind` support is "solved" until we both remove the `-Cpanic=abort`, and the behavior reaches a stable release of rustc.

PiperOrigin-RevId: 663349913
Change-Id: I1ab8cb1131c038a023fceae123927eb22c1d0a0f