Rename `--bindings-from-dependency` to `--crate-header`, with an alias to keep old calls working.
The problem with the old name is that it doesn't generalize: it's specifying the bindings that were generated from a given dependency, but we also want to have per-crate arguments that _aren't_ generated "from" them, and that aren't only for dependencies. In particular, we'll want a per-crate crubit feature flag argument, where one of the crates is the current crate. `feature-from-dependency` is consistent with the existing naming scheme, but misleading, I feel.

So the choices were:

1. introduce a _new_ naming scheme for the per-crate feature flags, such as `crate-feature`, which is inconsistent with the per-crate header flag.
2. be consistent (`feature-from-dependency`), but misleading.
3. be consistent (`feature-from-dependency`), and make it not-as-misleading by using a different flag for the current non-dependency crate.

One last thing is that I think `V_from_K` or `V_of_K` etc. names for K->V maps are somewhat rare. For example, rs_bindings_from_cc names the equivalent map just `V_K`: `target_args`. So, here, we can name it just `crate-header`. It's not entirely self-descriptive, but it doesn't need to be.

"When should we delete the old name"? This isn't use in GNI or anything, so we can remove it in the next toolchain release.

PiperOrigin-RevId: 663419411
Change-Id: I0bd9acb503352b9cbd6a8e7925186ade3b652365