Collect no eligible ranges for types of implicit Decls.
This is intended to fix a narrow case of function type arguments in templated types for lambda-captured variables. This situation leads to incompletely initialized FunctionTypeLocs that caused crashes when we expected non-null ParmVarDecls to be found in the FunctionTypeLoc's params array.

In general, we can't write any annotation for a type of an implicit Decl, though, so we fix this with a more general change.

PiperOrigin-RevId: 665830491
Change-Id: Ia746f715d7ee800723074901cc92b1f1b6bd4c1c