[nullability] Change range reported for invalid `->` dereference.
Currently, the source range that is reported for an invalid `->` dereference
encompasses the pointer expression that is potentially null. Some systems,
however, report only the beginning of the range; for chained dereferences such
as `p1->p2->field`, this makes it ambiguous which of the two dereferences is the
issue (and which of the pointers is potentially null).

We change this to instead report the source range of the `->` operator, which
makes it unambiguous where the issue lies.

PiperOrigin-RevId: 666219792
Change-Id: Iabd46ed95430c24e25ccb680c48cc40085ef3194