Stop trying to walk partially lost TypeLocs for noreturn-ed aliases.
When a noreturn attribute is applied to a Typedef for a function pointer, the Type is unwrapped, modified to be noreturn, and then best-effort rewrapped, but type source information is lost in the case of Typedefs. Because the underlying type is now different, we can't simply rewrap in the same Typedef.

PiperOrigin-RevId: 667576397
Change-Id: If79df2c764337fe8e98d0b0188c305b3e112d353