Use a toolchain to resolve the crubit binary.
Without using toolchains, or some other solution that also spares the build graph to the same extent, C++/Rust interop via Crubit is incompatible with some binaries, because it would introduce an unacceptable increase to the size of the build graph at analysis time.

Before this CL, we used `select()` to find the actual binary that generates the bindings. `select()` is resolved pretty late, and so puts everything from every branch on the dependency graph at analysis-time. Yet production builds literally only take the branch of the `select()` which depends on just a single binary! So, by making this a toolchain, and evaluated sooner (as I understand it), we cut the number of imagined dependencies significantly.

(I am not a bazel expert.)

We can repeat this for the other uses of select in Crubit (ESPECIALLY for the C++ standard library) to cut the dependencies even further, hopefully so that, by the end of the toolchainification, all or almost all analysis-time deps are in fact the real dependencies.

---

This CL does not port over the prebuilt-for-debugging option. That was obsoleted by crubit on demand, so there's no point. It is now deleted.

I'm not 100% sure on code organization here, but overall I'm pretty pleased with how this worked out. I do think it will not be as pretty once we add standard library headers to this (possibly a bunch of toolchains created by a list comprehension), but on the plus side it might mean we can totally kill the is_on_demand bool, so we'll see.

PiperOrigin-RevId: 668800654
Change-Id: I36a5d8c3bff8c77cf15b1abdc07f75b7f9f27055