Move GRTE toolchain definitions out of cc_std, into common.
The problem is that in google, we branch cc_std together with libc++ into a new directory, so that they come together as one atomic unit. (That way, for instance, if std::string_view changes in some incompatible way, so can its supporting Rust code.) The toolchain definitions do _not_ need to be branched in this way, and when they are, the resulting toolchain definitions are dead code and totally unused. The toolchains we do register only refer to the copy at head in crubit/support/cc_std.

So, instead, they go together with all the other toolchain definitions in common/bazel_support/BUILD.

This is basically a cleanup change, and should not affect any behavior. In particular, the toolchain registrations themselves (in common/bazel_support/toolchains/) are just changed to refer to toolchain _definitions_ in common/bazel_support/, which are textually identical.

PiperOrigin-RevId: 674240330
Change-Id: I4974d818b606b6910ce8b735fa12e33fc178e955