Support `use` for types from extern crates.
For symbols from extern crates, we use the "extern path" as the canonical names because they are guarantee to accessible in the current crate. To do so, we need to check all the paths when we compute the canonical names. Consider the following example:
```
mod x {
   pub use extern_crate::Y;
}

pub use x::Y;
pub fn g() -> extern_crate::Z {}
```
We have three paths here: `extern_crate::Y`, `x::Y` and `extern_crate::Z` but `x::Y` is not an "extern path" because `x` is a local mod. Therefore, the canonical name of `Y` will be `extern_crate::Y`.

Ideally we want to compute the canonical names for each crate and reuse them but I haven't found a way to access the items from extern crates.

PiperOrigin-RevId: 676030634
Change-Id: I8417d83ecbcbec1a7736290d30ea5cc346c2e3c4