Implement Crubit feature requirements for types.
Well, there's only one right now, but it's structured so that we can do more later.

This is the second time I'm implementing this, and the second implementation is much cleaner: to track the error messages, etc., I just use a fine grained enum. That's it! Doing something with string values is a bad idea. :(

Deployment plan for this CL (and any other CLs that gate to experimental):

- [x] wait until a version of crubit that accepts feature flags is released to the stable toolchain.
- [x] change the bzl to implicitly pass in `:experimental` to all targets -- done in https://github.com/google/crubit/commit/1fd4aec28f57ae402cb5ce32d4af6e8eec446d66
- [ ] it's now safe to submit gating CLs, including this one.
- [ ] once we're done gating features, apply `:experimental` _manually_ to all targets that need it, and submit a bzl change that reverts #2. (Which targets to apply `:experimental` to, can be determined by the breakage list from the CL reverting #2.)

PiperOrigin-RevId: 676632613
Change-Id: I693be7e380bfefef3db898a9f02512c678148fd9