Allow `clone_from` to be called from non-experimental users, despite taking a `&mut` and `&`.
This is safe, because `operator=` actually checks for aliasing before calling clone_from. A typical `operator=` impl looks something like:

```c++
inline T& T::operator=(const T& other) {
  if (this != &other) {
    __crubit_internal::
        __crubit_thunk__blahblahblah_clone_from_blahblahblah(*this, other);
  }
  return *this;
}
```

This kind of aliasing comparison is exactly what we want to enable in b/351876244, but in this case, it's already been hardcoded, so we should just grant this one special case permission.

PiperOrigin-RevId: 678013977
Change-Id: If0f0d8d4e213b6344237c5f16e15210c50370df6