Add MacOS and Windows build instructions for the Cargo build
MacOS needs -lc++, Windows does not need any explicit library
included. MacOS needs to be pointed to the XCode sdk, at least
if using a custom-built compiler, so include it in the steps.
Windows needs to be pointed to the Zlib headers and library that
were used to build LLVM.

PiperOrigin-RevId: 679673564
Change-Id: Ic7465c2e364203bf10283af703d59423d4d5422c