Safely allow multiple const references in Rust functions called from C++.
If any reference is mutable, we still require **exactly** one lifetime in play, out of the fear that mutable aliasing may occur if called from C++. But for non-mut references, it's fine to have interfaces like `fn foo(_: &i32, _: &i32) {}`.

**Future work:** Without any runtime checks, we could also later extend this to the following parameter types:

* Const references which share a lifetime should be OK, it's just a little annoying to write the code, so I didn't do it at first. The following should be allowed: `fn foo<'a>(_: &'a i32, _: &'a i32)`. No UB is possible from calling this from C++.
* Nullable references, such as as `Option<&i32>`, or `Option<&mut i32>`, should be just as allowed as non-optional references. This requires a bit more digging into the parameter types, but is a reasonable extension with no downsides really.

b/351876244 tracks this overall project, as well as adding runtime checks so that we can have `&mut` references mixed with non-`mut` references, or even have references as public struct fields, etc., and it will be checked for non-aliasing at runtime.

PiperOrigin-RevId: 679778884
Change-Id: Iae07052b2af3f8b00e0741637e9e3e7348c5a26a